Move `MainNavbarContainer` to its own directory
