+++ frontend/src/metabase/nav/containers/MainNavbar/MainNavbarContainer/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./MainNavbarContainer";
